Refactored from main to their respective django apps
