+++ authentication/templates/login.html
@@ -38,7 +38,7 @@
   </ul>
   {% endif %}
 
-  Don't have an account yet? <a href="{% url 'main:register' %}">Register Now</a>
+  Don't have an account yet? <a href="{% url 'authentication:register' %}">Register Now</a>
 
 </div>
