+++ calorie_tracking/templates/calorie_tracking.html
@@ -8,12 +8,12 @@
 
 <body>
   <h1>Food is not your enemy!</h1>
-  <a href="{% url 'main:problem_report' %}">
+  <a href="{% url 'problem_report:problem_report' %}">
     <button>
       Problem Report
     </button>
   </a>
-  <a href="{% url 'main:logout' %}">
+  <a href="{% url 'authentication:logout' %}">
     <button>
       Logout
     </button>
