+++ calorie_tracking/templates/stats.html
@@ -1,0 +1,57 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <meta charset="UTF-8">
+  <title>Stats</title>
+</head>
+
+<body>
+  <h1>Tell us about you</h1>
+  <a href="{% url 'problem_report:problem_report' %}">
+    <button>
+      Problem Report
+    </button>
+  </a>
+  <a href="{% url 'authentication:logout' %}">
+    <button>
+      Logout
+    </button>
+  </a>
+
+  <div>
+    <label for="weight">Weight (kg):</label>
+    <input type="number" id="weight" name="weight">
+  </div>
+  <div>
+    <label for="age">Age:</label>
+    <input type="number" id="age" name="age">
+  </div>
+  <div>
+    <label for="height">Height (cm):</label>
+    <input type="number" id="height" name="height">
+  </div>
+  
+  <button id="calculate-button">Calculate Recommended Calories</button>
+
+  <p id="recommended-calories"></p>
+
+  <script>
+    document.getElementById("calculate-button").addEventListener("click", function () {
+      const weight = parseFloat(document.getElementById("weight").value);
+      const age = parseInt(document.getElementById("age").value);
+      const height = parseFloat(document.getElementById("height").value);
+
+      if (isNaN(weight) || isNaN(age) || isNaN(height)) {
+        alert("Please enter valid numbers for weight, age, and height.");
+        return;
+      }
+
+      const recommendedCalories = Math.round(10 * weight + 6.25 * height - 5 * age + 5);
+      document.getElementById("recommended-calories").textContent = `Recommended Calories per Day: ${recommendedCalories}`;
+    });
+  </script>
+
+</body>
+
+</html>
