+++ list_information/templates/list.html
@@ -11,10 +11,10 @@
   <h1>Welcome to Body'N'Soul!</h1>
 
   <div>Don't have an account yet?
-    <a href="{% url 'main:register' %}">Register Now</a>
+    <a href="{% url 'authentication:register' %}">Register Now</a>
   </div>
   <div>Are you already a member?
-    <a href="{% url 'main:login' %}">Login Now</a>
+    <a href="{% url 'authentication:login' %}">Login Now</a>
   </div>
   <table border="1">
     <thead>
